bpf-mons: allow target less mode
We usually require either a target PID to attach to
or exec command to spawn and target process.  In some
cases, however, we need to allow target-less mode, when
bpf-mon monitors system wide events.

BUG=b:443191504
TEST=compile tested

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6919074
Commit-Queue: Sergey Senozhatsky <senozhatsky@chromium.org>
Tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>

+++ bpf-mons/libmon.cc
@@ -76,6 +76,10 @@ int setup_sig_handlers(void) {
 int prepare_target(pid_t& pid, const char* cmd, std::vector<char*>& args) {
   pid_t npid = -1;
 
+  if (pid == -1 && args.empty()) {
+    return 0;
+  }
+
   if (pid != -1) {
     kill(pid, SIGSTOP);
     return 0;
@@ -114,6 +118,9 @@ int prepare_target(pid_t& pid, const char* cmd, std::vector<char*>& args) {
 }
 
 int follow_target(pid_t pid) {
+  if (pid == -1) {
+    return 0;
+  }
   /* Setup is done now we can resume target process */
   return kill(pid, SIGCONT);
 }
